# Conflicts:
#	client/components/fields/Activities/Edit.jsx
#	client/components/fields/Activities/List.jsx
#	client/components/fields/Activities/View.jsx
#	client/components/getField.jsx
#	nscreg.Data/Configuration/ActivityConfiguration.cs
#	nscreg.Data/Entities/StatisticalUnit.cs
#	nscreg.Data/Migrations/20170322090535_Initial.Designer.cs
#	nscreg.Data/Migrations/20170322090535_Initial.cs
#	nscreg.Data/Migrations/20170328131016_Initial.Designer.cs
#	nscreg.Data/Migrations/20170328131016_Initial.cs
#	nscreg.Data/Migrations/20170328140511_initial.Designer.cs
#	nscreg.Data/Migrations/20170328140511_initial.cs
#	nscreg.Data/Migrations/NSCRegDbContextModelSnapshot.cs
#	nscreg.Data/NSCRegDbContext.cs
#	nscreg.Data/NscRegDbInitializer.cs
#	nscreg.ModelGeneration/PropertyMetadataBase.cs
#	nscreg.Resources/Languages/Resource.Designer.cs
#	nscreg.Resources/Languages/Resource.ky-KG.resx
#	nscreg.Resources/Languages/Resource.resx
#	nscreg.Resources/Languages/Resource.ru-RU.resx
#	nscreg.Server.Test/ActivityServiceTest.cs
#	nscreg.Server.Test/StatUnitServiceTest.cs
#	nscreg.Server/Controllers/Activities.cs
#	nscreg.Server/Models/AutoMapperConfiguration.cs
#	nscreg.Server/Models/StatUnits/ActivityM.cs
#	nscreg.Server/Services/ActivityService.cs
#	nscreg.Server/Services/StatUnitService.cs
#	nscreg.Server/Validators/StatUnitCreateEditBaseValidator.cs
#	nscreg.Server/wwwroot/dist/assets.json
#	nscreg.Server/wwwroot/dist/main.js
#	nscreg.Server/wwwroot/dist/main.js.map
#	nscreg.Utilities/ObjectComparer.cs

+++ nscreg.CommandStack/CommandContext.cs
@@ -31,11 +31,14 @@ namespace nscreg.CommandStack
             _dbContext.SaveChanges();
         }
 
-        public async Task SuspendRole(string id)
+        public async Task ToggleSuspendRole(string id, RoleStatuses status)
         {
-            _dbContext.Roles.Single(x => x.Id == id).Status = RoleStatuses.Suspended;
-            var records = await _dbContext.UserRoles.Where(x => x.RoleId == id).ToListAsync();
-            _dbContext.UserRoles.RemoveRange(records);
+            _dbContext.Roles.Single(x => x.Id == id).Status = status;
+            if (status == RoleStatuses.Suspended)
+            {
+                var records = await _dbContext.UserRoles.Where(x => x.RoleId == id).ToListAsync();
+                _dbContext.UserRoles.RemoveRange(records);
+            }
             await _dbContext.SaveChangesAsync();
         }
 
@@ -43,15 +46,14 @@ namespace nscreg.CommandStack
 
         #region USERS
 
-        public async Task SetUserStatus (string id, UserStatuses status)
+        public async Task SetUserStatus(string id, UserStatuses status)
         {
             var user = _dbContext.Users.Single(x => x.Id == id);
             user.Status = status;
-            user.SuspensionDate = status == UserStatuses.Suspended ? DateTime.Now : (DateTime?)null;
+            user.SuspensionDate = status == UserStatuses.Suspended ? DateTime.Now : (DateTime?) null;
             await _dbContext.SaveChangesAsync();
         }
 
         #endregion
-
     }
 }
